finished the tutorial, now let's explore it

+++ app/Http/Controllers/PostController.php
@@ -7,12 +7,24 @@ use Illuminate\Http\Request;
 
 class PostController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(['auth'])->only(['store', 'destroy']);
+    }
+
     public function index()
     {
         $posts = Post::latest()->with(['user', 'likes'])->paginate(20);
 
         return view('posts.index', [
             'posts' => $posts
+        ]);
+    }
+
+    public function show(Post $post)
+    {
+        return view('posts.show', [
+            'post' => $post
         ]);
     }
 
@@ -34,9 +46,11 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
-        if (!$post->ownedBy(auth()->user())) {
-            dd('ok');
-        }
+        /* if (!$post->ownedBy(auth()->user())) {
+            dd('no');
+        } */
+
+        $this->authorize('delete', $post);
 
         $post->delete();
